\name{RQTLNetwork-package}
\alias{RQTLNetwork-package}
\alias{RQTLNetwork}
\docType{package}
\title{Make Graphics for QTLNetwork}
\description{RQTLNetwork package is made for improve QTLNetwork's }
\details{
\tabular{ll}{
Package: \tab RQTLNetwork\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-06-07\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
type RQTLNetwork() in R, and there's a gWidgets GUI shows
}
\author{
Zheng,Wenjun
Maintainer: Zheng,Wenjun <wjzh@zju.edu.cn>
}
\references{
QTLNetwork
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{
R
QTLNetwork
QTL mapping
}
\examples{
RQTLNetwork()
}
